pyproject.toml
[project]
name = "mygallery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt>=0.9",
    "easyocr>=1.7.2",
    "faiss-cpu==1.11.0",
    "google-search-results==2.4.2",
    "langchain==0.3.25",
    "langchain-community==0.3.23",
    "langchain-core==0.3.58",
    "langchain-faiss==0.1.1",
    "langchain-openai==0.3.16",
    "langchain-text-splitters==0.3.8",
    "openai==1.77.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "pydantic==2.11.4",
    "pypdf==5.4.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv==1.1.0",
    "python-pptx>=1.0.2",
    "tavily-python==0.7.1",
    "watchdog==6.0.0",
]
